Extract Postgres helpers
